pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "support-agent-openenv"
version = "0.1.0"
description = "An OpenEnv SaaS Customer Support Agent Environment"
authors = [
    {name = "Shibashis Das"}
]
requires-python = ">=3.10"
dependencies = [
    "openenv-core",
    "pydantic",
    "fastapi",
    "uvicorn",
    "openai"
]

# Required by OpenEnv validator
[project.scripts]
server = "server.app:main"